Bill S4726A version A - James Sanders Jr.
BILL NUMBER: S4726A SPONSOR: SANDERS TITLE OF BILL: An act to amend the tax law, in relation to permitting a potential lottery winner who lost their ticket to claim the winning amount PURPOSE OR GENERAL IDEA OF BILL: This Bill would allow a potential lottery winner to claim the winning amount by providing proof of purchase of the winning lottery ticket. SUMMARY OF PROVISIONS: Section 1 amends section 1614 of the tax law to add "lost lottery ticket claims" to the title of the section.