Bill S6820B version B - Pete Harckham
BILL NUMBER: S6820B SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the public health law and the agriculture and markets law, in relation to requiring allergen labeling for prepackaged foods PURPOSE: Requires allergen labeling for prepackaged foods. SUMMARY OF PROVISIONS: Section 1 amends the public health law by adding a new section 1357 which will require allergen labeling on all food that is prepared, prepacked, and offered or sold to customers on the same premises.