Bill S8115B version B - James Skoufis
BILL NUMBER: S8115B SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the tax law, in relation to the timing of annual tax elections PURPOSE: This bill would amend the tax law to allow a taxpayer to make the PTET election by September 15 of the tax year to which the election applies. Current law requires the election to be made by March 15 of the tax year.