Bill S8824 version original - Jack M. Martins
BILL NUMBER: S8824 SPONSOR: MARTINS TITLE OF BILL: CONCURRENT RESOLUTION OF THE SENATE AND ASSEMBLY proposing the addition of a new subdivision to section 5-b of article 3 of the constitution, in relation to limiting the power of the legislature to amend a redistrict- ing plan submitted by the independent redistricting commission PURPOSE: Limits the power of the legislature to amend a redistricting plan submitted by the independent redistricting commission. SUMMARY OF PROVISIONS: Section one.