Bill S8840 version original - Anthony H. Palumbo
BILL NUMBER: S8840 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the retirement and social security law, in relation to the optional twenty-five year retirement of constables employed by the village of Westhampton Dunes PURPOSE: Relates to the optional twenty-five-year retirement of constables employed by the village of Westhampton Dunes. SUMMARY OF PROVISIONS: Section 1.