Bill S8837 version original - Leroy Comrie
BILL NUMBER: S8837REVISED 3/19/24 SPONSOR: COMRIE TITLE OF BILL: An act to amend the public health law, in relation to temporary services provided by "health care personnel" PURPOSE OF BILL: To clarify the use of temporary services within the definition of healthcare workforce personnel used in the context of the temporary staffing registry. SUMMARY OF PROVISIONS OF BILL: This bill amends section 2999-ii (4.