Bill S8842 version original - Anthony H. Palumbo
BILL NUMBER: S8842 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the labor law, in relation to the prohibition on retali- atory action by employers against whistleblowers; and to amend chapter 522 of the laws of 2021 amending the labor law relating to retaliatory actions by employers, in relation to the effectiveness of such provisions PURPOSE: To prohibit retaliatory action against certain employees. SUMMARY OF PROVISIONS: Section 1.