Bill S3170C version B - James Skoufis
BILL NUMBER: S3170B SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the domestic relations law and the family court act, in relation to establishing "Kyra's Law" PURPOSE OR GENERAL IDEA OF BILL: To protect children by ensuring courts assess serious risks to their safety in child custody and visitation proceedings. SUMMARY OF PROVISIONS: Section 1 of the bill identifies the name of this legislation as "Kyra's Law.