Bill S3141A version A - John Mannion
BILL NUMBER: S3141A SPONSOR: MANNION TITLE OF BILL: An act to amend the private housing finance law, in relation to the mobile and manufactured home replacement program PURPOSE: The purpose of this bill is to establish the mobile and manufactured home replacement program. SUMMARY OF PROVISIONS: Section 1 of this bill would amend the Private Housing Finance Law by establishing the Mobile and Manufactured Home Replacement program.