Bill S7086B version A - Pete Harckham
BILL NUMBER: S7086A SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the environmental conservation law and the state finance law, in relation to the disposition of certain fees and penalties PURPOSE: Relates to the disposition of certain fees and penalties. SUMMARY OF PROVISIONS: Section 1 amends the environmental conservation law to establish that fees paid for certain environmental violations will be deposited into New York's conservation enforcement account.