Bill S8844 version original - Sean Ryan
BILL NUMBER: S8844 SPONSOR: RYAN TITLE OF BILL: An act to amend the state finance law, in relation to adding employee- owned enterprises to the list of preferred contractors for public contracts in the state PURPOSE: To strengthen the viability and economic opportunity of employee-owned enterprises. SUMMARY OF PROVISIONS: Section 1: defines an employee-owned enterprise.