Bill S8847 version original - Jamaal Bailey
BILL NUMBER: S8847 SPONSOR: BAILEY TITLE OF BILL: An act to amend the tax law and the alcoholic beverage control law, in relation to the suspension or revocation of certain licenses for violations of the cannabis law PURPOSE: This bill would permit the revocation of a tobacco, liquor or gaming license of a business found to have sold or possessed illicit cannabis. SUMMARY OF PROVISIONS: Section 1.