Bill S6796A version A - Joseph P. Addabbo Jr.
BILL NUMBER: S6796A SPONSOR: ADDABBO TITLE OF BILL: An act to amend the agriculture and markets law, in relation to strengthening enforcement efforts of the state's anti-slaughter of race horses and race horse breeding stock law PURPOSE: To amend the agriculture and markets law, in relation to strengthening enforcement efforts of the state's anti-slaughter of racehorses and racehorse breeding stock law. SUMMARY OF PROVISIONS: Section 1.