Bill S8821A version A - Luis R. Sepúlveda
BILL NUMBER: S8821A SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the education law, the correction law, and the adminis- trative code of the city of New York, in relation to changing the name of the general education development diploma to the Excelsior high school diploma PURPOSE: This bill directs the Commissioner of Education to change the name of the New York state high school equivalency diploma to the Excelsior High School Diploma. SUMMARY OF PROVISIONS: Section 1.