Bill S4680A version A - Kevin S. Parker
BILL NUMBER: S4680A SPONSOR: PARKER TITLE OF BILL: An act to amend the criminal procedure law, in relation to establishing an automatic expunction system for certain records PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to address issues related to the practice of "stop and frisk" by NYPD and to remedy the constitutional violations found in Floyd v. City of New York.