Bill S8857 version original - Monica Martinez
BILL NUMBER: S8857 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to increasing the penalties for stopping, standing, or parking in handicapped spaces without a handicapped permit PURPOSE: This legislation will amend the vehicle and traffic law to increase the penalties for stopping, standing or parking in handicapped spaces with- out a handicapped permit. SUMMARY OF PROVISIONS: Section 1.