Bill S2689C version C - Neil D. Breslin
BILL NUMBER: S2689C SPONSOR: BRESLIN TITLE OF BILL: An act to amend the executive law, the public authorities law and the public buildings law, in relation to the utilization of renewable energy at state-owned facilities in Albany PURPOSE: The bill requires that state facilities at the Empire State Plaza complex in downtown Albany shall entirely use renewables for power, heat and cooling. SUMMARY OF PROVISIONS: Section 1 states that the bill shall be known as the renewable Capitol Act.