Bill S8878 version original - James Sanders Jr.
BILL NUMBER: S8878 SPONSOR: SANDERS TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to violations for concealed license plates PURPOSE: To further discourage the purposeful obstruction of license plates. SUMMARY OF PROVISIONS: to increase fines for glass or plastic, or any conceals or obscures the image of the plate.