Bill S8859 version original - Monica Martinez
BILL NUMBER: S8859 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the penal law, in relation to assault in the second degree of an operator or crew of a passenger commuter ferry PURPOSE: This legislation would amend the penal law to include the intentional causing of physical injury to a passenger commuter ferry operator or crew as a Class D felony of assault in the second degree. SUMMARY OF PROVISIONS: Section 1. Subdivision 11 of section 120.