Bill S8868 version original - Mario Mattera
BILL NUMBER: S8868 SPONSOR: MATTERA TITLE OF BILL: An act to amend the real property actions and proceedings law, in relation to providing an exclusion for treble damages in actions for forcible or unlawful entry or detainer PURPOSE: Section 853 of the real property actions and proceedings law provides for an action for unlawful entry or detainer. This legislation would exclude from the definition of "person" squatters and trespassers.