Bill S8876 version original - Leroy Comrie
BILL NUMBER: S8876 SPONSOR: COMRIE TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to increasing the number of licenses for tow truck operators PURPOSE: This bill would authorize New York City to increase the number of licenses for tow truck operators. SUMMARY OF PROVISIONS: Section 1 adds subparagraph b to Section 20-520 of the administrative code of the city of New York to increase available tow truck licenses by thirty percent.