Bill S8890 version original - Neil D. Breslin
BILL NUMBER: S8890 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the general business law, in relation to establishing standards for glass repair and calibration of advanced driver assistance systems PURPOSE: The purpose of this bill is to provide important consumer protections for the replacement and recalibration of automotive safety glass. SUMMARY OF PROVISIONS: Section 1 includes new requirements for motor vehicle glass repair and replacement.