Bill S7090A version A - Leroy Comrie
BILL NUMBER: S7090A SPONSOR: COMRIE TITLE OF BILL: An act to amend the public authorities law, in relation to providing notification to customers of bed bug infestations on MTA subways, trains and buses PURPOSE OR GENERAL IDEA OF BILL: To notify MTA customers, making them aware of bed bug infestations on MTA subways, trains, and buses as a measure of public safety and public health awareness, to help prevent the spreading of such infestation. SUMMARY OF SPECIFIC PROVISIONS: Section 1.