Bill S8892A version original - Jamaal Bailey
BILL NUMBER: S8892 SPONSOR: BAILEY TITLE OF BILL: An act to amend the executive law, in relation to directing the office of children and family services to establish a center for fatherhood initiatives PURPOSE: This bill requires the office of children and family services to estab- lish a center for fatherhood initiatives where entities can apply for grants to implement programs that support positive fatherhood in the state of New York. SUMMARY OF PROVISIONS: Section 1.