Bill S8889 version original - Neil D. Breslin
BILL NUMBER: S8889 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to loss prevention programs PURPOSE OF THE BILL: The purpose of this bill is to encourage property insurers to offer voluntary loss mitigation programs to their insureds. SUMMARY OF PROVISION: Section 2324 of the Insurance Law is amended to add a new subsection 6 to allow insurers to provide loss mitigation programs to policyholders pursuant to regulations promulgated by the superintendent.