Bill S8883 version original - Jeremy Cooney
BILL NUMBER: S8883 SPONSOR: COONEY TITLE OF BILL: An act to amend chapter 109 of the laws of 2018 amending the general business law relating to certain rental vehicle protections, in relation to making such provisions permanent SUMMARY OF PROVISIONS: Section 1 amends the effective date of Chapter 163 of the Laws of 2023 to make permanent the provisions of Chapter 656 of the Laws of 2002 as amended by Chapter 109 of the Laws of 2018. Section 2 is the effective date.