Bill S8880A version original - James Sanders Jr.
BILL NUMBER: S8880 SPONSOR: SANDERS TITLE OF BILL: An act to amend the general business law, in relation to the sale of over-the-counter drugs and cosmetic products PURPOSE: To promote and protect the health and safety of purchasers of O-T-C drugs and cosmetics by enabling them to read such products' packaging or labels in a complete, unaltered and unobstructed condition, at the time of purchase or thereafter. Such labels are required by federal law and regulations.