Bill S8886 version original - Rachel May
BILL NUMBER: S8886 SPONSOR: MAY TITLE OF BILL: An act to amend the environmental conservation law, in relation to updating the definition of the term "reusable bag" PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to update the definition of a "reuseable bag" to include bags made from plastic that can be recycled. SUMMARY OF PROVISIONS: Section 1 expands the definition of "reuseable bags." Section 2 sets the effective date.