Bill S8904 version original - Iwen Chu
BILL NUMBER: S8904 SPONSOR: CHU TITLE OF BILL: An act to amend the penal law, in relation to repeated instances of harassment PURPOSE: The purpose of this legislation is to ensure repeat offenders are charged in accordance with the compounded results of their recidivism. SUMMARY OF PROVISIONS: Section 1 amends subdivision 5 of section 240.30 of the penal law to include instances of repeat harassment as a charge of aggravated harass- ment in the second degree and harassment in the first degree.