Bill S6798A version A - Peter Oberacker
BILL NUMBER: S6798A SPONSOR: OBERACKER TITLE OF BILL: An act to amend the criminal procedure law, in relation to permitting certain defendants in a criminal action in Delaware, Otsego or Schoharie county to appear electronically, with the approval of the court PURPOSE: This bill will permit the use of virtual arraignments, at the court's discretion, in Delaware, Otsego, and Schoharie counties. SUMMARY OF PROVISIONS: Section one of this bill amends section 182.