Bill S8903 version original - Monica Martinez
BILL NUMBER: S8903 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the real property tax law, in relation to certain real property tax exemptions to include additional eligibility for people with disabilities PURPOSE: The purpose of the bill is to expand accessibility to partial real prop- erty tax exemptions for persons with disabilities. SUMMARY OF PROVISIONS: Section 1.