Bill S8900A version original - Andrea Stewart-Cousins
BILL NUMBER: S8900 SPONSOR: STEWART-COUSINS TITLE OF BILL: An act to amend chapter 62 of the laws of 2015 amending the tax law relating to the imposition of an occupancy tax in the city of Yonkers, in relation to extending the effectiveness thereof PURPOSE: Extends the effectiveness of an occupancy tax imposed in the City of Yonkers. SUMMARY OF PROVISIONS: Amends Chapter 112 of the Laws of 2018 extending the sunset from Septem- ber 1, 2024 to September 1, 2027.