Bill S8915 version original - Monica Martinez
BILL NUMBER: S8915 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the general municipal law and the banking law, in relation to the establishment of a savings bank, savings and loan asso- ciation or credit union municipal deposit program PURPOSE OR GENERAL IDEA OF BILL: To expand the options for local governments to choose who they bank with to include credit unions, savings banks, and savings and loan associ- ations. SUMMARY OF PROVISIONS: Section 1.