Bill S8916 version original - Neil D. Breslin
BILL NUMBER: S8916 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the executive law, in relation to providing for certain victim statements to be taken at their workplace instead of the police department PURPOSE OR GENERAL IDEA OF BILL: To permit health care workers who have been assaulted on the job to give their statements to the police at their workplace. SUMMARY OF PROVISIONS: Section 1 adds a new subdivision 6 to section 642 of the executive law. Section 2 provides an effective date.