Bill S8926 version original - Pete Harckham
BILL NUMBER: S8926 SPONSOR: HARCKHAM TITLE OF BILL: An act in relation to directing the public service commission to create an agrivoltaics incentive program PURPOSE: Directs the public service commission to create an agrivoltaics incen- tive program. SUMMARY OF PROVISIONS: Section 1 states the legislative findings and statement of purpose, which is to meet the CLCPA target for 70% renewable energy in NY by 2030, in addition to providing financial benefits for agricultural lands.