Bill S6967A version A - Cordell Cleare
BILL NUMBER: S6967A SPONSOR: CLEARE TITLE OF BILL: An act to amend the banking law, the estates, powers and trusts law, the public health law and the real property law, in relation to providing for the remote conduct of certain practices and procedures PURPOSE: To modernize various provisions in state law relating to remote prac- tices and procedures. Codifies Executive Orders of 2020: 80 202.5, EC 202.7, E0 202.8, BO 202.13, BO 202.14, EC 202.15, and EC 202.18.