Bill S8942 version original - Monica Martinez
BILL NUMBER: S8942 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the executive law, in relation to the establishment of local comprehensive emergency management plans PURPOSE: The purpose of this legislation is to ensure local emergency management plans include animal accommodation preparation and public communication that considers the needs of persons with vision or hearing impairments. SUMMARY OF PROVISIONS: Section 1.