Bill S8944 version original - Kevin S. Parker
BILL NUMBER: S8944 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to the definition of autism PURPOSE OR GENERAL IDEA OF BILL: To insert into the mental hygiene law a definition of autism and other autism spectrum disorders. The definition is the current generally accepted definition as set forth in the American Psychiatric Associ- ation's Diagnostic and Statistical Manual of Mental Disorders (DSM). SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends Section 1.