Bill S8413A version A - Neil D. Breslin
BILL NUMBER: S8413A SPONSOR: BRESLIN TITLE OF BILL: An act to amend the penal law, in relation to establishing the crime of staging a construction site accident PURPOSE OR GENERAL IDEA OF BILL: This critical piece of legislation is designed to crack down on a grow- ing and pernicious form of insurance fraud: staged accidents on construction sites. SUMMARY OF PROVISIONS: Section 1 amends section 176 of the penal law to make staging a construction.