Bill S8945 version original - Neil D. Breslin
BILL NUMBER: S8945 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the public officers law, in relation to the residency requirement for assistant district attorneys in Schenectady county PURPOSE: Permits residents of Schenectady County and adjoining counties to hold the position of assistant district attorney in Schenectady County. SUMMARY OF PROVISIONS: Section 1.