Bill S8949A version original - Jessica Ramos
BILL NUMBER: S8949 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to establishing a workplace readiness week for secondary school students PURPOSE: This bill would require the Department of Labor to provide a document on workplace rights to any minor seeking working papers. SUMMARY OF PROVISIONS: Section 1: Requires that the Department of Labor provides a "Workplace Rights Document" to any minor seeking working papers in order to be employed in the state.