Bill S8954 version original - Cordell Cleare
BILL NUMBER: S8954 SPONSOR: CLEARE TITLE OF BILL: An act to amend the real property tax law, in relation to increasing the amount of income property owners may earn for the purpose of eligibility for the property tax exemption for persons sixty-five years of age or over PURPOSE OR GENERAL IDEA OF BILL: To increase the maximum income allowable for senior citizen homeowners to be eligible for a tax exemption SUMMARY OF PROVISIONS: Section 1 increases. the maximum income allowable to $75,000.