Bill S8958 version original - Michelle Hinchey
BILL NUMBER: S8958 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the general municipal law, in relation to authorizing regional planning council's eligibility to apply for federal and/or state grants PURPOSE: To allow for Regional Planning Councils the eligibility of state and federal grants. SUMMARY OF PROVISIONS: Amends the general municipal law to allow Regional Planning Councils the eligibility of state and federal grants.