Bill S8968 version original - Monica Martinez
BILL NUMBER: S8968 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the vehicle and traffic law and the penal law, in relation to regulating the purchase and sale of catalytic converters PURPOSE: The purpose of this legislation is to clarify who can buy or sell used, detached catalytic converters. This legislation further prohibits the possession of two or more catalytic converters by any individual who is not registered as a scrap metal dealer, vehicle dismantler, or repair shop.