Bill S8974 version original - Jessica Scarcella-Spanton
BILL NUMBER: S8974 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the environmental conservation law, in relation to the burning of wood, coal, natural gas, propane, or other fuels for the purposes of cooking or otherwise preparing food SUMMARY OF PROVISIONS: Section 1. Establishes the title for the function as "Preserving Our Culinary Traditions Act". Section 2.