Bill S8975 version original - Jessica Scarcella-Spanton
BILL NUMBER: S8975 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to exempting health care workers from congestion pricing SUMMARY OF PROVISIONS: Section 1. Amends Section 1704-a of the vehicle and traffic law by stip- ulating that the Triborough Bridge and Tunnel Authority shall implement a plan to provide credits or exemptions for tolls issued to health-care workers on bridges and crossings. Section 2.