Bill S8985 version original - Jessica Scarcella-Spanton
BILL NUMBER: S8985 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the retirement and social security law, in relation to military service credit SUMMARY OF PROVISIONS: Section 1. Amends Subdivision 1 of section 1000 of the retirement and social security law to clarify the definition of a qualifying condition and a discharged LGBTQ+ veteran. It also includes active military duty to include services performed under the authority of the Governor of New York. Section 2.