Bill S8998 version original - Monica Martinez
BILL NUMBER: S8998 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the executive law, in relation to the timing of emergen- cy public safety alerts in the event of a disaster or state disaster emergency PURPOSE: The purpose of this legislation is to ensure NY-Alerts are issued in a timely manner to inform and secure the public in the event of a disas- ter. SUMMARY OF PROVISIONS: Section 1.