Bill S8990 version original - Jeremy Cooney
BILL NUMBER: S8990 SPONSOR: COONEY TITLE OF BILL: An act to amend the retirement and social security law, in relation to prior police service with an accredited police agency outside of the state of New York SUMMARY OF PROVISIONS: Section 1 adds a new paragraph 3 to subdivision c of section 381-b of the retirement and social security law to allow for a service credit of up to three years of out of state service at an accredited police agen- cy. Section 2 is the effective date.