Bill S9012 version original - Monica Martinez
BILL NUMBER: S9012 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the public health law, in relation to the sale, manufac- ture, and distribution of vapor products PURPOSE: The purpose of this legislation is to strengthen regulations and crite- ria related to the sale, manufacture, and distribution of vapor products. SUMMARY OF PROVISIONS: Section 1.