Bill S9006 version original - Kevin S. Parker
BILL NUMBER: S9006 SPONSOR: PARKER TITLE OF BILL: An act to amend the energy law, in relation to the creation of geother- mal facilities in New York state PURPOSE OR GENERAL IDEA OF BILL: AN ACT to amend the energy law, in relation to the creation of geother- mal facilities in New York state. SUMMARY OF PROVISIONS: 1. The energy law is amended by adding a new section 3-105. Geothermal facilities.