Bill S9008 version original - Jessica Scarcella-Spanton
BILL NUMBER: S9008 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the social services law, in relation to payments of certain expenses by adoptive parents SUMMARY OF PROVISIONS: Section 1. Amends Subdivision 6 of section 374 of the social services law by adding subdivision 6(b) language that extends the period in which payments to a birthmother shall be made by an adoptive family. Section 2.