Bill S9010 version original - James Skoufis
BILL NUMBER: S9010 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the public officers law, in relation to expanding use of videoconferencing to purely advisory bodies PURPOSE: This bill is intended to eliminate physical presence quorum requirements in the Open Meetings Law for advisory committees and subcommittees of public bodies. SUMMARY OF PROVISIONS: Section 1.