Bill S9026 version original - Anthony H. Palumbo
BILL NUMBER: S9026 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the town law, in relation to permitting employees of a fire department to serve as fire chief or assistant fire chief PURPOSE: To allow volunteer firefighters to hold a dual position. SUMMARY OF PROVISIONS: Section 1. Subdivision 11-a of section 176 of the town law, as amended by chapter 351 of the laws of 2000 is amended to allow for nominees to include employees of the district. Section 2.