Bill S4941A version A - Kevin S. Parker
BILL NUMBER: S4941A SPONSOR: PARKER TITLE OF BILL: An act to amend the energy law and the public authorities law, in relation to the availability of alternative fuels along the New York state thruway PURPOSE OR GENERAL IDEA OF BILL: The intent of this bill is to make alternative fuels and electric charg- ing stations available to commuters who use the thruway. SUMMARY OF SPECIFIC PROVISIONS: Section 1: This act shall be known and may be cited as the "electric super highway act of 2023".