Bill S9024 version original - Sean Ryan
BILL NUMBER: S9024 SPONSOR: RYAN TITLE OF BILL: An act to amend the judiciary law, in relation to requiring town and village courts compensate temporary interpreters for deaf or hard of hearing persons PURPOSE: The per diem rate for temporary interpreters for deaf or hard of hearing persons at town and village courts has been set at $25 for decades. This bill seeks to increase the compensation provided to such interpreters.