Bill S9034 version original - Jeremy Cooney
BILL NUMBER: S9034 SPONSOR: COONEY TITLE OF BILL: An act to amend the criminal procedure law, in relation to electronic appearance in connection with a criminal action pending in Monroe county SUMMARY OF PROVISIONS: Section one of the bill amends section 182.20 of the criminal procedure law to include Monroe County. Section two is the effective date.