Bill S9017 version original - Samra Brouk
BILL NUMBER: S9017 SPONSOR: BROUK TITLE OF BILL: An act to amend the mental hygiene law, in relation to disclosure of the records of court proceedings PURPOSE: This legislation would limit access to the court records in Article 81 Guardianship proceedings. SUMMARY OF PROVISIONS: Section 1. Amends subdivision. (b) of Section 81.14 of the Mental Hygiene Law to provide that the court records in a proceeding under Article 81.