Bill S3587A version A - Pamela Helming
BILL NUMBER: S3587A SPONSOR: HELMING TITLE OF BILL: An act to amend the mental hygiene law, in relation to establishing the rural suicide prevention council; and providing for the repeal of such provisions upon expiration thereof PURPOSE: The purpose of this bill is to reduce the growing rate of suicide in rural New York-through the establishment of a rural suicide prevention council. SUMMARY OF PROVISIONS: Creates a new section 7.