Bill S8388B version B - Pete Harckham
BILL NUMBER: S8388B SPONSOR: HARCKHAM TITLE OF BILL: An act in relation to protecting the health insurance benefits of reti- rees of public employers and contributions of retirees of public employ- ers PURPOSE OR GENERAL IDEA OF BILL: Prohibits public employers from diminishing the health insurance bene- fits of public employee retirees. SUMMARY OF PROVISIONS: Section 1 establishes that the name of the act shall be the "Health Equity for Retirees Act.