Bill S7623C version B - Brad Hoylman-Sigal
BILL NUMBER: S7623B SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the labor law, in relation to restricting the use of electronic monitoring and automated employment decision tools; and to amend the civil rights law, in relation to making a conforming change PURPOSE: To regulate the use of electronic monitoring and automated decision tools in employment SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds a new section article 36 to the labor law. New section 1010 of article 36 defines terms.