Bill S9050 version original - Michelle Hinchey
BILL NUMBER: S9050 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the real property tax law, in relation to real property tax exemptions for veterans PURPOSE: To provide a real property tax exemption to veterans who have served on submarines or on classified missions during a period of war. SUMMARY OF PROVISIONS: Section 1 provides for a 10 percent property tax abatement for veterans who have served on classified missions or on a submarine during a period of war.