Bill S1608A version A - José M. Serrano
BILL NUMBER: S1608A SPONSOR: SERRANO TITLE OF BILL: An act to amend the parks, recreation and historic preservation law and the general municipal law, in relation to farmers' markets in state and local parks PURPOSE: The purpose of this legislation is to increase the number of available locations for farmers' markets and increase New Yorker's access to locally grown and produced food. SUMMARY OF PROVISIONS: Section 1 - Amends section 3.