Bill S6752B version B - Pete Harckham
BILL NUMBER: S6752B SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the education law, in relation to local block anesthesia PURPOSE: This bill allows hygienists to administer block anesthesia in addition to the currently allowed local anesthesia and infiltration, in conform- ance with regulations of the commissioner. SUMMARY OF PROVISIONS: This bill amends article 133, sections 6606 and 6605-b, of the education law.