Bill S385A version A - George Borrello
BILL NUMBER: S385A SPONSOR: BORRELLO TITLE OF BILL: An act to amend the criminal procedure law, in relation to peace officer status of public safety officers of the Chautauqua Institution within the grounds of and properties owned by the Chautauqua Institution PURPOSE OR GENERAL IDEA OF BILL: This bill would designate, as peace officers, public safety officers of the Chautauqua Institution within the grounds of and properties owned by the Chautauqua Institution. SUMMARY OF PROVISIONS: Section 1.