Bill S7053A version A - Brad Hoylman-Sigal
BILL NUMBER: S7053A SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the penal law, in relation to including the abduction of a child to avoid treatment for gender dysphoria in the offense of custo- dial interference in the first degree PURPOSE OR GENERAL IDEA OF BILL: To strengthen the New York State penal law of custodial interference in the first degree. SUMMARY OF PROVISIONS: Section 1. Amends penal law 135.