Bill S1114A version A - George Borrello
BILL NUMBER: S1114A SPONSOR: BORRELLO TITLE OF BILL: An act to amend the criminal procedure law, in relation to designating uniformed court officers in the town of Busti, county of Chautauqua, as peace officers PURPOSE OR GENERAL IDEA OF BILL: This bill would designate the uniformed court officers of the Town of Busti, as peace officers. SUMMARY OF PROVISIONS: Section 1. Amends Section 2.