Bill S1997A version A - Rachel May
BILL NUMBER: S1997A SPONSOR: MAY TITLE OF BILL: An act to amend the labor law, in relation to enacting the "standing is tiring (sit) act" PURPOSE: To enhance workplace safety and performance by allowing employees who are reasonably able to sit during their jobs the option to do so. SUMMARY OF PROVISIONS: Section 1: Sets name for this act as the Standing Is Tiring (SIT) Act.