Bill S9098 version original - Luis R. Sepúlveda
BILL NUMBER: S9098 SPONSOR: SEPULVEDA PURPOSE: Clarifies the definition of "tenant" in the real property actions and proceeding law to exclude "squatters" and provides for expedited adjudi- cation by the courts. SUMMARY OF PROVISIONS: Section one: The opening paragraph of section 711 of the real property actions and proceedings law is amended to define the term "tenant" to exclude "squatter." such tenant must have been in possession for ninety consecutive days or longer.