Bill S9089 version original - Rachel May
BILL NUMBER: S9089 SPONSOR: MAY TITLE OF BILL: Relates to qualification for holding the office of code enforcement officer of the City of Auburn. PURPOSE: To facilitate the hiring of code enforcement personnel in the City of Auburn. SUMMARY OF PROVISIONS: Section 1: adds a new subdivision to Section 3 of the public officers law to allow the city of Auburn to hire code enforcement officers who live in Cayuga County or adjoining counties. Section 4: Sets the effective date.