Bill S9090 version original - Sean Ryan
BILL NUMBER: S9090 SPONSOR: RYAN TITLE OF BILL: An act to amend the private housing finance law, the tax law and the real property tax law, in relation to enacting the accessory dwelling unit incentive act PURPOSE: The purpose of this bill is to establish financial incentives to build accessory dwelling units as a means of creating more affordable housing. SUMMARY OF PROVISIONS: Section 1: Short title.