Bill S9101 version original - Monica Martinez
BILL NUMBER: S9101 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend chapter 18 of the laws of 2020 authorizing the commis- sioner of education to appoint a monitor to oversee the Wyandanch union free school district and establishing the powers and duties of the moni- tor, in relation to extending the powers of the monitor PURPOSE: The purpose of this iezislation is to grant the Wyandanch Union Free School District Fiscal Monitor veto powers. SUMMARY OF PROVISONS: Section 1.