Bill S9123B version original - Kristen Gonzalez
BILL NUMBER: S9123 SPONSOR: GONZALEZ TITLE OF BILL: An act to amend the multiple dwelling law, in relation to coverage of a unit or occupant and the status of an interim multiple dwelling unit as a protected occupant's primary residence PURPOSE OR GENERAL IDEA OF BILL: To clarify tenant protections under the Loft Law. SUMMARY OF PROVISIONS: Section 1 sets forth legislative findings.