Bill S9133A version original - Jeremy Cooney
BILL NUMBER: S9133 SPONSOR: COONEY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to allowing for individuals to enroll in the national bone marrow donor program at the time of application for or renewal of a driver's license or non-driver identification card PURPOSE: To allow anyone who is applying for or renewing a driver's license the option to enroll in the national bone marrow donor program. Section 1.