Bill S9137A version original - Michael Gianaris
BILL NUMBER: S9137 SPONSOR: GIANARIS TITLE OF BILL: An act to amend the penal law, in relation to openly carrying a rifle or shotgun PURPOSE: This bill aims to regulate the open carry of rifles and shotguns with enumerated exceptions in order to promote public safety and align with existing law in other states. SUMMARY OF PROVISIONS: Section 1. Amends penal law 265.