Bill S9141 version original - Brad Hoylman-Sigal
BILL NUMBER: S9141 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the civil rights law, in relation to ensuring the priva- cy of name change actions and of sex designation change actions; and to repeal certain provisions of such law relating thereto PURPOSE: To ensure the privacy of legal documents relating to name changes. SUMMARY OF PROVISIONS: Section 1 repeals section 64-a of the civil rights law and adds a new section 64-a governing the privacy of name change actions.