Bill S9149A version original - Nathalia Fernandez
BILL NUMBER: S9149 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the public health law and the administrative code of the city of New York, in relation to prohibiting charges for the issuance of a certificate of still birth or fetal death PURPOSE OR GENERAL IDEA OF BILL: This bill prohibits fees associated with obtaining a certificate of stillbirth or fetal death. SUMMARY OF PROVISIONS: Section 1.