Bill S9157 version original - Anthony H. Palumbo
BILL NUMBER: S9157 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the judiciary law, in relation to production of appel- late briefs PURPOSE: Provides public access to appellate briefs. SUMMARY OF PROVISIONS: Section 1 amends the judiciary law by adding a new section 432-a stating the court of appeals shall provide a copy of any appellate brief filed with court upon request to the clerk of the chief judge.