Bill S9150 version original - Brad Hoylman-Sigal
BILL NUMBER: S9150 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the public health law, in relation to prohibiting the sale of oral nicotine pouches PURPOSE: To prohibit the sale of flavored oral nicotine pouches. SUMMARY OF PROVISIONS: Section 1. Amends section 1399-aa of the public health law to add a new subdivision 19. This adds the definition of nicotine pouch. Section 2. Amends section 1399-bb of the public health law to amend section 4 of part EE of chapter 56 of the laws of 2020.