Bill S9177 version original - Anthony H. Palumbo
BILL NUMBER: S9177 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the public health law, in relation to requests for birth certificates PURPOSE: Relates to requests for birth certificates. SUMMARY OF PROVISIONS: Section 1. Amends subdivision 2 of the section 4173 of he public health law, relating to the issuance of a certified copy or certified tran- script of a birth record. Section 2. Sets for the effective date.