Bill S9182 version original - Toby Ann Stavisky
BILL NUMBER: S9182 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to the professional certification of doulas PURPOSE: To provide the requirements for obtaining the title of "certified doula". SUMMARY OF PROVISIONS: Section 1: Sets forth and defines applicable definitions, and outlines the requirements necessary for obtaining the title of "certified doula". Section 2: States the effective date.