Bill S9183B version original - Anthony H. Palumbo
BILL NUMBER: S9183 SPONSOR: PALUMBO TITLE OF BILL: An act in relation to authorizing the trustees of the Tuckahoe Common School District, county of Suffolk, to submit a proposition to the qual- ified electors of such district to reorganize such school district as a union free school district PURPOSE: To propose that the Tuckahoe Common School District be reorganized as a union free school district. SUMMARY OF PROVISIONS: Section 1.