Bill S9196 version original - Toby Ann Stavisky
BILL NUMBER: S9196 SPONSOR: STAVISKY TITLE OF BILL: An act in relation to authorizing the commissioner of education to conduct an audit regarding instruction in literacy instruction within the state PURPOSE: Authorizes and directs the Commissioner of Education to conduct a program audit in regard to instruction in literacy at institutions of higher education in the state. SUMMARY OF PROVISIONS: Section 1.