Bill S9200 version original - Joseph A. Griffo
BILL NUMBER: S9200 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the legislative law, in relation to establishing a legislative commission on article V conventions of the states PURPOSE: Allows the legislature to set procedures for selecting, instructing and overseeing commissioners for an Article V Convention of States. SUMMARY OF PROVISIONS: 1.