Bill S9191 version original - Robert Rolison
BILL NUMBER: S9191 SPONSOR: ROLISON TITLE OF BILL: An act to amend the executive law, in relation to adding Fishkill Creek to the definition of inland waterways PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to designate the Fishkill Creek and its major tributaries as an inland waterway. SUMMARY OF PROVISIONS: Section 1: Amends paragraph (c) of subdivision 4 of section 911 of the Executive Law by adding Fishkill Creek to the list of the state's major creeks.