Bill S1865B version B - Samra Brouk
BILL NUMBER: S1865B SPONSOR: BROUK TITLE OF BILL: An act to amend the education law, in relation to requiring an education campaign about the 9-8-8 suicide and crisis lifeline and the crisis text line PURPOSE: Require institutions of higher education to establish an education campaign to educate campus communities about the 9-8-8 suicide and crisis lifeline and the crisis text line. SUMMARY OF PROVISIONS: Section 1: Provides title of the act.