Bill S9194 version original - Steve Rhoads
BILL NUMBER: S9194 SPONSOR: RHOADS TITLE OF BILL: An act to amend the penal law, in relation to establishing the crimes of deceptive wearing of a mask and aggravated deceptive wearing of a mask PURPOSE OR GENERAL IDEA OF BILL: This bill addresses the growing concern surrounding the deceptive wear- ing of masks during public protests, rallies, and assemblies. SUMMARY OF PROVISIONS: Section 1 amends the penal law by adding two new sections: 190.90 adds the crime of deceptive mask wearing; 190.