Bill S9209 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S9209 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the regu- lation of the operation of motorized scooters PURPOSE: To codify all motorized scooters in law and promote their safe opera- tion. SUMMARY OF PROVISIONS: Section 1 amends the vehicle and traffic law by adding a new section 123-a to define motorized scooters in statute.