Bill S9218 version original - Jack M. Martins
BILL NUMBER: S9218 SPONSOR: MARTINS TITLE OF BILL: An act to amend the Nassau county civil divisions act, in relation to the name and the funds of the corporation PURPOSE OF THE BILL: This bill will amend chapter 73 of the laws of 1940 relating to the incorporating of the Alert Volunteer Exempt Fireman's Benevolent Associ- ation of Great Neck, New York, Inc. SUMMARY OF PROVISIONS: Section one of the bill amends section 614.0 and 614.