Bill S9222 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S9222 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the racing, pari-mutuel wagering and breeding law, in relation to the creation of the jockey health insurance reserve fund PURPOSE OR GENERAL IDEA OF BILL: Authorizes the establishment of an additional segregated account for the creation of a reserve fund to pay for the payment of premiums not yet paid. SUMMARY OF PROVISIONS: Section 1 of this bill amends section 221-a of the racing, pari-mutuel wagering and breeding law.