Bill S9213 version original - Toby Ann Stavisky
BILL NUMBER: S9213 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to physical therapist assistants PURPOSE: To provide greater efficiency and flexibility in regards to the duties of physical therapist assistants. SUMMARY OF PROVISIONS: Section 1: Outlines the duties of the physical therapist assistant. Section 2: Sets forth the effective date.