Bill S4788A version A - Gustavo Rivera
BILL NUMBER: S4788A SPONSOR: RIVERA TITLE OF BILL: An act to amend the social services law, in relation to the use of transportation management brokers PURPOSE: Provides reimbursement of certain transportation costs under Medicaid. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends subparagraph (vi) of paragraph (b) of subdivision 4 of section 365-h of the social services law as it relates to the reimburse- ment of transportation costs for medicaid enrollees.