Bill S9234 version original - Liz Krueger
BILL NUMBER: S9234 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the real property law, in relation to prohibiting rent minimums in mortgages PURPOSE OR GENERAL IDEA OF BILL: This bill would prohibit mortgagees from penalizing or defaulting on mortgagors for not charging a high enough rent. Eliminating the minimum rent requirement for mortgagors will incentivize landlords to rent prop- erties at a lower rate and not risk defaulting.