Bill S7721A version A - Peter Oberacker
BILL NUMBER: S7721A SPONSOR: OBERACKER TITLE OF BILL: An act to amend the navigation law, in relation to regulations to restrict docks, boathouses and moorings on Otsego Lake in the town of Springfield PURPOSE: To allow the local legislative body in the town of Springfield in the County of Otsego to adopt, amend and enforce local laws, rules and regu- lations not inconsistent with the laws of the United States. EXISTING LAW: Section 1.